feat(shared): notebook updates
