+++ libs/shared/src/components/file/file-editor.tsx
@@ -3,8 +3,8 @@ import { FileCodeEditor } from "./file-code-editor";
 import { FileDownloadView } from "./file-download-view";
 import { FileImageViewer } from "./file-image-viewer";
 import { FileMarkdownEditor } from "./file-markdown-editor";
+import { FileNotebook } from "./file-notebook";
 import { FilePdfViewer } from "./file-pdf-viewer";
-import { Notebook } from "./notebook";
 
 // Extensions that cannot be rendered in the editor — show a download-first view instead
 const NON_RENDERED_EXTENSIONS = new Set([
@@ -86,7 +86,7 @@ export const FileEditor: React.FC<FileEditorProps> = ({
 			)}
 			{/* .ipynb → interactive notebook renderer/runner */}
 			{isNotebook && (
-				<Notebook
+				<FileNotebook
 					key={path}
 					mode={mode}
 					path={path}
